Controlladores tutorial

+++ 02-Acme_Antenna/Workspace/Acme-Antenna/src/main/java/services/TutorialService.java
@@ -8,10 +8,13 @@ import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.Validator;
 
 import repositories.TutorialRepository;
 import domain.Actor;
 import domain.Tutorial;
+import forms.TutorialForm;
 
 @Service
 @Transactional
@@ -26,6 +29,12 @@ public class TutorialService {
 
 	@Autowired
 	private ActorService actorService;
+	
+	@Autowired
+	private AdministratorService administratorService;
+	
+	@Autowired
+	private Validator validator;
 
 	// Constructors
 
@@ -40,17 +49,11 @@ public class TutorialService {
 		res = new Tutorial();
 
 		Date moment = new Date(System.currentTimeMillis() - 1000);
-		String title = "title";
-		String text = "text";
-		String pictures = "http://www.google.es";
 
 		Actor actor;
 		actor = this.actorService.findByPrincipal();
 
 		res.setMoment(moment);
-		res.setText(text);
-		res.setTitle(title);
-		res.setPictures(pictures);
 		res.setActor(actor);
 
 		return res;
@@ -81,11 +84,47 @@ public class TutorialService {
 	}
 
 	public void delete(Tutorial tutorial) {
+		administratorService.checkAuthority();
 		Assert.notNull(tutorial);
 		Assert.isTrue(tutorial.getId() != 0);
 		Assert.isTrue(this.tutorialRepository.exists(tutorial.getId()));
 		this.tutorialRepository.delete(tutorial);
 	}
 
+	
 	// Other business method --------------------------------------------------
+	
+	public TutorialForm construct(Tutorial tutorial) {
+		Assert.notNull(tutorial);
+		TutorialForm res = new TutorialForm();
+		
+		res.setId(tutorial.getId());
+		res.setTitle(tutorial.getTitle());
+		res.setText(tutorial.getText());
+		res.setPictures(tutorial.getPictures());
+		res.setActorId(tutorial.getActor().getId());
+		
+		return res;
+	}
+
+	public Tutorial reconstruct(TutorialForm tutorialForm,
+			BindingResult binding){
+		Assert.notNull(tutorialForm);
+		Tutorial res;
+//		Date moment = new Date(System.currentTimeMillis() - 1000);
+		
+		if(tutorialForm.getId()!=0)
+			res = this.findOne(tutorialForm.getId());
+		else
+			res = this.create();
+		
+		res.setTitle(tutorialForm.getTitle());
+		res.setText(tutorialForm.getText());
+		res.setPictures(tutorialForm.getPictures());
+		
+		if(binding!=null)
+			this.validator.validate(res,binding);
+		
+		return res;
+	}
 }
